move account loading logic to route logic
