+++ packages/web/src/pages/sign-up-page/SignUpPage.tsx
@@ -1,12 +1,11 @@
 import { Helmet } from 'react-helmet'
 import { Redirect, Route, RouteProps, Switch } from 'react-router-dom'
 
-import { getSignOn } from 'common/store/pages/signon/selectors'
-import { determineAllowedRoute } from 'pages/sign-up-page/utils/determineAllowedRoutes'
-import { useSelector } from 'utils/reducer'
+import { useDetermineAllowedRoute } from 'pages/sign-up-page/utils/useDetermineAllowedRoutes'
 import {
   SIGN_UP_APP_CTA_PAGE,
   SIGN_UP_ARTISTS_PAGE,
+  SIGN_UP_COMPLETED_REDIRECT,
   SIGN_UP_CREATE_LOGIN_DETAILS,
   SIGN_UP_EMAIL_PAGE,
   SIGN_UP_FINISH_PROFILE_PAGE,
@@ -15,7 +14,8 @@ import {
   SIGN_UP_LOADING_PAGE,
   SIGN_UP_PAGE,
   SIGN_UP_PASSWORD_PAGE,
-  SIGN_UP_REVIEW_HANDLE_PAGE
+  SIGN_UP_REVIEW_HANDLE_PAGE,
+  TRENDING_PAGE
 } from 'utils/route'
 
 import { CreateEmailPage } from './pages/CreateEmailPage'
@@ -39,7 +39,7 @@ const messages = {
  * <Route> wrapper that handles redirecting through the sign up page flow
  */
 export function SignUpRoute({ children, ...rest }: RouteProps) {
-  const signUpState = useSelector(getSignOn)
+  const determineAllowedRoute = useDetermineAllowedRoute()
 
   return (
     <Route
@@ -47,7 +47,6 @@ export function SignUpRoute({ children, ...rest }: RouteProps) {
       render={({ location }) => {
         // Check if the route is allowed, if not we redirect accordingly
         const { isAllowedRoute, correctedRoute } = determineAllowedRoute(
-          signUpState,
           location.pathname
         )
 
@@ -99,6 +98,9 @@ export const SignUpPage = () => {
         </SignUpRoute>
         <SignUpRoute exact path={SIGN_UP_LOADING_PAGE}>
           <LoadingAccountPage />
+        </SignUpRoute>
+        <SignUpRoute exact path={SIGN_UP_COMPLETED_REDIRECT}>
+          <Redirect to={TRENDING_PAGE} />
         </SignUpRoute>
       </Switch>
     </RouteContextProvider>
