+++ packages/web/src/pages/sign-up-page/pages/LoadingAccountPage.tsx
@@ -3,12 +3,11 @@ import { useEffect } from 'react'
 import { Flex, useTheme } from '@audius/harmony'
 import { useSelector } from 'react-redux'
 
-import { useModalState } from 'common/hooks/useModalState'
 import { getStatus } from 'common/store/pages/signon/selectors'
 import { EditingStatus } from 'common/store/pages/signon/types'
 import LoadingSpinner from 'components/loading-spinner/LoadingSpinner'
 import { useNavigateToPage } from 'hooks/useNavigateToPage'
-import { TRENDING_PAGE } from 'utils/route'
+import { SIGN_UP_COMPLETED_REDIRECT } from 'utils/route'
 
 import { Heading, Page } from '../components/layout'
 
@@ -17,10 +16,9 @@ const messages = {
   description: "We're just finishing up a few things..."
 }
 
-// This loading page shows up when the user's account is still being created either due to slow creation or a fast user
+// This loading page shows up when the users account is still being created either due to slow creation or a fast user
 // The user just waits here until the account is created and before being shown the welcome modal on the trending page
 export const LoadingAccountPage = () => {
-  const [, setIsWelcomeModalOpen] = useModalState('Welcome')
   const navigate = useNavigateToPage()
   const { color } = useTheme()
 
@@ -28,11 +26,10 @@ export const LoadingAccountPage = () => {
 
   useEffect(() => {
     if (accountCreationStatus === EditingStatus.SUCCESS) {
-      navigate(TRENDING_PAGE)
-      setIsWelcomeModalOpen(true)
+      navigate(SIGN_UP_COMPLETED_REDIRECT)
     }
     // TODO: what to do in an error scenario? Any way to recover to a valid step?
-  }, [navigate, setIsWelcomeModalOpen, accountCreationStatus])
+  }, [navigate, accountCreationStatus])
 
   return (
     <Page gap='3xl' justifyContent='center' alignItems='center' pb='3xl'>
