+++ packages/web/src/pages/sign-up-page/pages/SelectArtistsPage.tsx
@@ -18,18 +18,12 @@ import { range } from 'lodash'
 import { useDispatch } from 'react-redux'
 import { toFormikValidationSchema } from 'zod-formik-adapter'
 
-import { useModalState } from 'common/hooks/useModalState'
 import { addFollowArtists } from 'common/store/pages/signon/actions'
-import { getGenres, getStatus } from 'common/store/pages/signon/selectors'
-import { EditingStatus } from 'common/store/pages/signon/types'
+import { getGenres } from 'common/store/pages/signon/selectors'
 import { useMedia } from 'hooks/useMedia'
 import { useNavigateToPage } from 'hooks/useNavigateToPage'
 import { useSelector } from 'utils/reducer'
-import {
-  SIGN_UP_APP_CTA_PAGE,
-  SIGN_UP_LOADING_PAGE,
-  TRENDING_PAGE
-} from 'utils/route'
+import { SIGN_UP_APP_CTA_PAGE, SIGN_UP_COMPLETED_REDIRECT } from 'utils/route'
 
 import { AccountHeader } from '../components/AccountHeader'
 import {
@@ -57,7 +51,6 @@ const initialValues: SelectArtistsValues = {
 
 export const SelectArtistsPage = () => {
   const artistGenres = useSelector((state) => ['Featured', ...getGenres(state)])
-  const [, setIsWelcomeModalOpen] = useModalState('Welcome')
   const [currentGenre, setCurrentGenre] = useState('Featured')
   const dispatch = useDispatch()
   const navigate = useNavigateToPage()
@@ -69,24 +62,17 @@ export const SelectArtistsPage = () => {
     setCurrentGenre(e.target.value)
   }, [])
 
-  const accountCreationStatus = useSelector(getStatus)
-
   const handleSubmit = useCallback(
     (values: SelectArtistsValues) => {
       const { selectedArtists } = values
       dispatch(addFollowArtists([...selectedArtists]))
       if (isMobile) {
-        if (accountCreationStatus === EditingStatus.SUCCESS) {
-          navigate(TRENDING_PAGE)
-          setIsWelcomeModalOpen(true)
-        } else {
-          navigate(SIGN_UP_LOADING_PAGE)
-        }
+        navigate(SIGN_UP_COMPLETED_REDIRECT)
       } else {
         navigate(SIGN_UP_APP_CTA_PAGE)
       }
     },
-    [accountCreationStatus, dispatch, isMobile, navigate, setIsWelcomeModalOpen]
+    [dispatch, isMobile, navigate]
   )
 
   const isFeaturedArtists = currentGenre === 'Featured'
