+++ packages/web/src/utils/route.ts
@@ -97,7 +97,8 @@ export enum SignUpPath {
   selectGenres = 'select-genres',
   selectArtists = 'select-artists',
   loading = 'loading',
-  appCta = 'app-cta'
+  appCta = 'app-cta',
+  completedRedirect = 'completed'
 }
 export const SIGN_UP_EMAIL_PAGE = `/signup/${SignUpPath.createEmail}`
 export const SIGN_UP_START_PAGE = SIGN_UP_EMAIL_PAGE // entry point for sign up if needing to redirect to the beginning
@@ -110,6 +111,7 @@ export const SIGN_UP_GENRES_PAGE = `/signup/${SignUpPath.selectGenres}`
 export const SIGN_UP_ARTISTS_PAGE = `/signup/${SignUpPath.selectArtists}`
 export const SIGN_UP_APP_CTA_PAGE = `/signup/${SignUpPath.appCta}`
 export const SIGN_UP_LOADING_PAGE = `/signup/${SignUpPath.loading}`
+export const SIGN_UP_COMPLETED_REDIRECT = `/signup/${SignUpPath.completedRedirect}`
 
 // Param routes.
 export const NOTIFICATION_USERS_PAGE = '/notification/:notificationId/users'
